[진행중] 장바구니 담기 기능 구현
1. Cart, CartItem 저장 로직 구현
2. 요청 시 405 Method Not Allowed 에러 발생

+++ src/main/resources/db/H2/schema.sql
@@ -109,7 +109,7 @@ CREATE TABLE cart_item (
                        cart_id BIGINT NOT NULL,
                        menu_id BIGINT NOT NULL,
                        quantity INT NOT NULL,
-                       price DECIMAL(10,2),
+                       price INT,
                        FOREIGN KEY (cart_id) REFERENCES cart (id),
                        FOREIGN KEY (menu_id) REFERENCES menu (id)
 );
